agentic-healthcare-demo/backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.8.2
python-dotenv==1.0.1
httpx==0.27.0
langchain==0.2.14
langchain-core==0.2.34
langchain-community==0.2.12
langchain-openai==0.1.22
openai==1.40.6
